New sort button added.
Decided to remove the sort functionality from undo. Instead added a new button
for it. It is a little repetitive for the user to sort everytime there is an
undo, but it is unavoidable if we want to sort those lines arbitrarily where
user can add id of a line.