Added new functionality on hovering lines and refactoring
Added some new functionality concerining the hovering of the mouse within
canvas.
The idea is, as the mouse hovers the line bounding boxes would become visible.
The coordination of this with the rectangle selection is important.

Rectangle selection should deactivate the drawing of the bounding box.
Also did some refactoring on functions for restoring the old canvas state