Major refactoring using classes
Did a major refactoring using functional classes.

Now there are 2 classes:
1, is dealing with the canvas
2, is dealing with the transcription column.

When there is a need for exchanging objects, they exchange it in the html
functions scope, that is they exchange objects only when they are called from
html functions.

The idea is to keep transcriptions separated from images

Next things to implement on the order of importance:

Hovering lines on image: couple of checks don't function very well

Passing indexes to and from transcription to image lines:
The idea is to attribute each selection an id that would be used for adding
the transcription line to the transcription column.
We calculate the id for the next line that is going to be added to the
transcription column and use that as an id(index) for the ongoing selection.
If the selection is associated to a transcription box we proceed with the
regular adding of transcription box

Filebrowser: to upload texts from local sources

Maybe a polygon selector would be nice.