The interface is ready for beta testing.
All the functions of the interface is done.
Now we need to implement it to the kraken.
The idea is either we ship widgets.js with html page or we create an assets
folder to where we save the transcription page.

There are also two TODOs:

1. Include adding a file and associating it to the image
2. Once we change the layout.html template I need to correct the todo
in the last iteration for saving everything.