Started svg version of the image viewer. Continued on the canvas version
as well.

Started working on an svg version of the image viewer.
The idea is to make the widget more maintainable. Since svg is xml based
it should work on browsers that disable js by default.
Plus it is easier to work with svgs since we are dealing with layer like
modifications, canvas are not really good for retaining this type of layering